+++ sandpack-react/src/components/Preview/Preview.stories.tsx
@@ -7,8 +7,9 @@ import {
   SandpackProvider,
   SandpackLayout,
 } from "../../";
+import { useSandpack } from "../../hooks";
 
-import type { PreviewProps } from "./index";
+import type { PreviewProps, SandpackPreviewRef } from "./index";
 import { SandpackPreview } from "./index";
 
 export default {
@@ -122,3 +123,43 @@ export const AdditionalButtons: React.FC = () => (
     </SandpackLayout>
   </SandpackProvider>
 );
+
+export const AdditionalContent: React.FC = () => (
+  <SandpackProvider template="react">
+    <SandpackLayout>
+      <SandpackPreview>
+        <div style={{ background: "lightgreen" }}>content after iframe</div>
+      </SandpackPreview>
+      <SandpackCodeEditor />
+    </SandpackLayout>
+  </SandpackProvider>
+);
+
+const SandpackClient: React.FC = () => {
+  const { sandpack } = useSandpack();
+  const previewRef = React.useRef<SandpackPreviewRef>();
+
+  React.useEffect(() => {
+    const client = previewRef.current?.getClient();
+    const clientId = previewRef.current?.clientId;
+
+    if (client && clientId) {
+      /* eslint-disable no-console */
+      console.log(client);
+      console.log(sandpack.clients[clientId]);
+    }
+  }, [sandpack]);
+
+  return <SandpackPreview ref={previewRef} />;
+};
+
+export const GetClient: React.FC = () => {
+  return (
+    <SandpackProvider template="react">
+      <SandpackLayout>
+        <SandpackClient />
+        <SandpackCodeEditor />
+      </SandpackLayout>
+    </SandpackProvider>
+  );
+};